update: 디렉토리 수정
