+++ src/test/java/gift/e2e/WishTest.java
@@ -1,16 +1,18 @@
-package gift.service;
+package gift.e2e;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.http.HttpMethod.DELETE;
+import static org.springframework.http.HttpMethod.GET;
 import static org.springframework.http.HttpMethod.POST;
 import static org.springframework.http.HttpMethod.PUT;
-import static org.springframework.http.HttpStatus.FORBIDDEN;
+import static org.springframework.http.HttpStatus.ALREADY_REPORTED;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.OK;
 
 import gift.auth.Login;
-import gift.domain.User.CreateUser;
-import gift.domain.User.UpdateUser;
+import gift.domain.Product.CreateProduct;
+import gift.domain.Product.UpdateProduct;
+import gift.domain.Wish.createWish;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -24,7 +26,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class UserTest {
+class WishTest {
 
     @LocalServerPort
     private int port;
@@ -39,12 +41,10 @@ class UserTest {
 
     @BeforeEach
     public void setUp() {
-
         Login login = new Login("kakao1@kakao.com", "1234");
 
         HttpEntity<Login> requestEntity = new HttpEntity<>(login);
-        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/login",
-            POST,
+        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/login", POST,
             requestEntity, String.class);
 
         int startIndex = responseEntity.getBody().indexOf("\"token\":\"") + "\"token\":\"".length();
@@ -56,84 +56,80 @@ class UserTest {
     }
 
     @Test
-    @DisplayName("유저 수정")
-    public void UpdateUser() {
-        UpdateUser body = new UpdateUser("123456789");
-
+    @DisplayName("위시리스트 조회(위시리스트 없음)")
+    public void NotFoundGetWish() {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.setBearerAuth(token);
 
-        HttpEntity<Long> requestEntity = new HttpEntity(body, headers);
-        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/user/1",
-            PUT, requestEntity, String.class);
-
-        System.out.println(responseEntity);
-        assertThat(responseEntity.getStatusCode()).isEqualTo(OK);
-    }
-
-    @Test
-    @DisplayName("유저 수정(유저 없음)")
-    public void NotFoundUpdateUser() {
-        UpdateUser body = new UpdateUser("123456789");
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        headers.setBearerAuth(token);
-
-        HttpEntity<Long> requestEntity = new HttpEntity(body, headers);
-        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/user/0",
-            PUT, requestEntity, String.class);
-
-        System.out.println(responseEntity);
-        assertThat(responseEntity.getStatusCode()).isEqualTo(NOT_FOUND);
-    }
-
-    @Test
-    @DisplayName("유저 생성")
-    public void CreateUser() {
-        CreateUser body = new CreateUser("kakao10@kakao.com", "1234");
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        headers.setBearerAuth(token);
-
-        HttpEntity<Long> requestEntity = new HttpEntity(body, headers);
-        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/user",
-            POST, requestEntity, String.class);
-
-        System.out.println(responseEntity);
-        assertThat(responseEntity.getStatusCode()).isEqualTo(OK);
-    }
-
-//    병렬로 수행하면 오류나지만, 단독으로 수행하면 수행 됨
-    @Test
-    @DisplayName("유저 생성(유저 중복)")
-    public void DuplicateCreateUser() {
-        CreateUser body = new CreateUser("kakao1@kakao.com", "1234");
-
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        headers.setBearerAuth(token);
-
-        HttpEntity<Long> requestEntity = new HttpEntity(body, headers);
-        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/user",
-            POST,
+        HttpEntity<Long> requestEntity = new HttpEntity(null, headers);
+        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/wish/9999", GET,
             requestEntity, String.class);
 
         System.out.println(responseEntity);
-        assertThat(responseEntity.getStatusCode()).isEqualTo(FORBIDDEN);
+        assertThat(responseEntity.getStatusCode()).isEqualTo(NOT_FOUND);
     }
 
     @Test
-    @DisplayName("유저 삭제")
-    public void removeUser() {
+    @DisplayName("위시리스트 생성")
+    public void CreateWish() {
+        createWish body = new createWish(3L);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        headers.setBearerAuth(token);
+
+        HttpEntity<Long> requestEntity = new HttpEntity(body, headers);
+        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/wish", POST,
+            requestEntity, String.class);
+
+        System.out.println(responseEntity);
+        assertThat(responseEntity.getStatusCode()).isEqualTo(OK);
+    }
+
+    @Test
+    @DisplayName("위시리스트 생성(상품 없음)")
+    public void NotFoundCreateWish() {
+        createWish body = new createWish(100L);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        headers.setBearerAuth(token);
+
+        HttpEntity<Long> requestEntity = new HttpEntity(body, headers);
+        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/wish", POST,
+            requestEntity, String.class);
+
+        System.out.println(responseEntity);
+        assertThat(responseEntity.getStatusCode()).isEqualTo(NOT_FOUND);
+    }
+
+    @Test
+    @DisplayName("위시리스트 생성(이미 존재)")
+    public void AlreadyCreateWish() {
+        createWish body = new createWish(2L);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        headers.setBearerAuth(token);
+
+        HttpEntity<Long> requestEntity = new HttpEntity(body, headers);
+        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/wish", POST,
+            requestEntity, String.class);
+
+        System.out.println(responseEntity);
+        assertThat(responseEntity.getStatusCode()).isEqualTo(ALREADY_REPORTED);
+    }
+
+    @Test
+    @DisplayName("위시리스트 삭제")
+    public void removeWish() {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.setBearerAuth(token);
 
         HttpEntity<Long> requestEntity = new HttpEntity(null, headers);
-        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/user/2",
+        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/wish/1",
             DELETE, requestEntity, String.class);
 
         System.out.println(responseEntity);
@@ -141,18 +137,17 @@ class UserTest {
     }
 
     @Test
-    @DisplayName("유저 삭제(유저 없음)")
-    public void NotFoundRemoveUser() {
+    @DisplayName("위시리스트 삭제(위시 리스트에 없음)")
+    public void NotFoundRemoveWish() {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.setBearerAuth(token);
 
         HttpEntity<Long> requestEntity = new HttpEntity(null, headers);
-        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/user/9999",
+        ResponseEntity<String> responseEntity = restTemplate.exchange(url + port + "/api/wish/9999",
             DELETE, requestEntity, String.class);
 
         System.out.println(responseEntity);
         assertThat(responseEntity.getStatusCode()).isEqualTo(NOT_FOUND);
     }
-
 }
